set route no login and logged

+++ src/pages/Profile.tsx
@@ -8,35 +8,21 @@ import MainLayouts from "../layouts/MainLayouts";
 import { useHistory } from "react-router";
 
 const Profile = () => {
-  const [res, setRes] = useState<typeUser>();
+  const [res, setRes] = useState<any>({});
   const [form] = Form.useForm();
   const history = useHistory();
-
-  useEffect(() => {
-    showData();
-  });
-
   const accessToken = localStorage.getItem("accessToken");
 
-  const showData = async () => {
-    const res = await axios.post(
-      `http://localhost:4000/api/user/getuser`,
-      { accessToken: accessToken },
-      {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      }
-    );
-    setRes(res.data);
-    localStorage.setItem("id", res.data.id);
-    localStorage.setItem("username", res.data.username);
-    localStorage.setItem("firstname", res.data.firstname);
-    localStorage.setItem("lastname", res.data.lastname);
-    localStorage.setItem("email", res.data.email);
-    localStorage.setItem("tel", res.data.tel);
-    console.log(res.data);
-  };
+  useEffect(() => {
+    setRes({
+      id: localStorage.getItem("id"),
+      username: localStorage.getItem("username"),
+      firstname: localStorage.getItem("firstname"),
+      lastname: localStorage.getItem("lastname"),
+      email: localStorage.getItem("email"),
+      tel: localStorage.getItem("tel"),
+    });
+  }, [res.length]);
 
   const deleteUser = async () => {
     const id = localStorage.getItem("id");
@@ -58,34 +44,39 @@ const Profile = () => {
     });
   };
 
-  return (
-    <>
-      <MainLayouts page="5">
-        <h4 >ชื่อ: {res?.firstname}</h4>
-        <h4>นามสกุล: {res?.lastname}</h4>
-        <h4>username: {res?.username}</h4>
-        <h4>email: {res?.email}</h4>
-        <h4>เบอร์โทร: {res?.tel}</h4>
-        
-        <Row justify="center">
-          <Button
-            onClick={() => {
-              deleteUser();
-            }}
-            type="link"
-            icon={
-              <DeleteTwoTone
-                twoToneColor="#FF0000"
-                style={{ fontSize: "30px" }}
-              />
-            }
-          />
+  if (!accessToken) {
+    history.push("/login");
+    return null;
+  } else {
+    return (
+      <>
+        <MainLayouts page="5">
+          <h4>ชื่อ: {res?.firstname}</h4>
+          <h4>นามสกุล: {res?.lastname}</h4>
+          <h4>username: {res?.username}</h4>
+          <h4>email: {res?.email}</h4>
+          <h4>เบอร์โทร: {res?.tel}</h4>
 
-          <EditUser />
-        </Row>
-      </MainLayouts>
-    </>
-  );
+          <Row justify="center">
+            <Button
+              onClick={() => {
+                deleteUser();
+              }}
+              type="link"
+              icon={
+                <DeleteTwoTone
+                  twoToneColor="#FF0000"
+                  style={{ fontSize: "30px" }}
+                />
+              }
+            />
+
+            <EditUser />
+          </Row>
+        </MainLayouts>
+      </>
+    );
+  }
 };
 
 export default Profile;
